refactor: モジュール `Preset` と `PresetError` を統合
